Rename errhandle.h -> exceptions.h.
