+++ logd/src/apps/logd/main.cpp
@@ -1,7 +1,7 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include <logd/config_subscriber.h>
-#include <logd/errhandle.h>
+#include <logd/exceptions.h>
 #include <logd/forwarder.h>
 #include <logd/metrics.h>
 #include <logd/state_reporter.h>
