+++ logd/src/logd/exceptions.h
