+++ logd/src/logd/forwarder.cpp
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include "errhandle.h"
+#include "exceptions.h"
 #include "forwarder.h"
 #include "metrics.h"
 #include <vespa/vespalib/component/vtag.h>
