+++ logd/src/logd/watcher.cpp
@@ -1,7 +1,7 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "config_subscriber.h"
-#include "errhandle.h"
+#include "exceptions.h"
 #include "forwarder.h"
 #include "watcher.h"
 #include <vespa/vespalib/util/sig_catch.h>
